+++ src/nodelets/rgbdicp_odometry.cpp
@@ -74,9 +74,9 @@ public:
 		exactCloudSync_(0),
 		queueSize_(5),
 		scanCloudMaxPoints_(0),
-		scanVoxelSize_(0.0f),
+		scanVoxelSize_(0.0),
 		scanNormalK_(0),
-		scanNormalRadius_(0.0f)
+		scanNormalRadius_(0.0)
 	{
 	}
 
@@ -464,9 +464,9 @@ private:
 	message_filters::Synchronizer<MyExactCloudSyncPolicy> * exactCloudSync_;
 	int queueSize_;
 	int scanCloudMaxPoints_;
-	float scanVoxelSize_;
+	double scanVoxelSize_;
 	int scanNormalK_;
-	float scanNormalRadius_;
+	double scanNormalRadius_;
 };
 
 PLUGINLIB_EXPORT_CLASS(rtabmap_ros::RGBDICPOdometry, nodelet::Nodelet);
